Allows the client to send the server appearance changes.

+++ src/main/java/com/valenguard/server/game/GameConstants.java
@@ -4,6 +4,6 @@ public class GameConstants {
     public static final int TILE_SIZE = 16;
     public static final int TICKS_PER_SECOND = 20;
 
-    public static final int HUMAN_MAX_HEADS = 79;
-    public static final int HUMAN_MAX_BODIES = 59;
+    public static final short HUMAN_MAX_HEADS = 79;
+    public static final short HUMAN_MAX_BODIES = 59;
 }
